+++ app/src/main/java/com/hulkdx/moneymanager/ui/main/MainMvpView.java
@@ -15,5 +15,5 @@ public interface MainMvpView extends MvpView {
 
     void showTransactions(List<Transaction> transactions);
 
-    void setBalanceTextView(int amount);
+    void setBalanceTextView(float amount);
 }
